backend/requirements.txt
Flask==2.0.2
google-cloud-storage==1.42.3
firebase-admin==5.0.3
google-cloud-firestore==2.1.0
